+++ app/src/main/java/nextstep/shoppingcart/MainActivity.kt
@@ -8,6 +8,7 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
+import nextstep.shoppingcart.data.Product
 import nextstep.shoppingcart.ui.productlist.ProductListScreen
 import nextstep.shoppingcart.ui.theme.ShoppingCartTheme
