--- app/src/main/java/nextstep/shoppingcart/ProductInfo.kt
@@ -1,10 +1,0 @@
-package nextstep.shoppingcart
-
-import java.io.Serializable
-
-data class Product(
-    val id: Int,
-    val imageUrl: String,
-    val name: String,
-    val price: Int,
-): Serializable
