+++ app/src/main/java/nextstep/shoppingcart/data/Cart.kt
@@ -1,7 +1,5 @@
 package nextstep.shoppingcart.data
 
-import nextstep.shoppingcart.Product
-
 data class CartItem(
     val product: Product,
     val count: Int,
