+++ app/src/main/java/nextstep/shoppingcart/ui/detail/ProductDetailScreen.kt
@@ -8,7 +8,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
-import nextstep.shoppingcart.Product
+import nextstep.shoppingcart.data.Product
 import nextstep.shoppingcart.R
 import nextstep.shoppingcart.data.Cart
 import nextstep.shoppingcart.ui.component.ShoppingCartNavigationTopBar
